+++ src/androidTest/java/youmeee/co/jp/clipper/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package youmeee.co.jp.clippablelayout;
+package youmeee.co.jp.clipper;
 
 import android.content.Context;
 import androidx.test.InstrumentationRegistry;
