+++ src/main/java/youmeee/co/jp/clipper/CircleClipEntry.kt
@@ -1,4 +1,4 @@
-package youmeee.co.jp.clippablelayout
+package youmeee.co.jp.clipper
 
 import android.graphics.Canvas
 import android.graphics.Paint
