+++ src/main/java/youmeee/co/jp/clipper/ClippableLayout.kt
@@ -1,4 +1,4 @@
-package youmeee.co.jp.clippablelayout
+package youmeee.co.jp.clipper
 
 import android.content.Context
 import android.view.View
