+++ src/main/java/youmeee/co/jp/clipper/ClippableQueueDispatcher.kt
@@ -1,4 +1,4 @@
-package youmeee.co.jp.clippablelayout
+package youmeee.co.jp.clipper
 
 interface ClippableQueueDispatcher {
     fun add(clippableLayout: ClippableLayout)
