+++ src/main/java/youmeee/co/jp/clipper/ClippableQueueDispatcherImpl.kt
@@ -1,4 +1,4 @@
-package youmeee.co.jp.clippablelayout
+package youmeee.co.jp.clipper
 
 import android.view.ViewGroup
 import android.view.Window
