+++ src/main/java/youmeee/co/jp/clipper/ClippableView.kt
@@ -1,4 +1,4 @@
-package youmeee.co.jp.clippablelayout
+package youmeee.co.jp.clipper
 
 import android.content.Context
 import android.graphics.*
